email-sender/requirements.txt
python-dotenv==1.0.0
sendgrid==6.11.0
pandas==2.2.0